+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v2.3.0] - 2022-07-27
+
+### Added
+
+- Add banner
+
+### Changed
+
+- Move session controller to separate file
+- Move header to separate file
+- Move footer to separate file
+
+### Fixed
+
+- Fix redirection in `index.php` form
+
 ## [v2.2.0] - 2022-04-19
 
 ### Added